refactor: TDDによるコンポーネント分解・ユーティリティ分割（Phase 1-5）
## Phase 1: analytics.ts ユーティリティ分割
- 519行のanalytics.tsを7つのモジュールに分割
- src/utils/analytics/: common, daily, weekly, monthly, predictions, aggregations, index
- 後方互換性のためのre-exportを維持

## Phase 2: ProjectCard コンポーネント分解
- 718行から228行に削減（68%減）
- useProjectCardState フック抽出
- ProjectProgressSection, ProjectCardActions コンポーネント作成
- 14件の新規テスト追加

## Phase 3: SettingsView モジュール化
- 488行から98行に削減（80%減）
- useSettingsState フック抽出
- WorkHoursSettings, AppearanceSettings, TestModeSettings コンポーネント作成

## Phase 4: TimerFocus コンポーネント分解
- 415行から199行に削減（52%減）
- useElapsedTime フック抽出
- TimerDisplay, TimerControls, NoteEditor コンポーネント作成

## Phase 5: ManualTimeEntryForm 状態管理改善
- 380行から134行に削減（65%減）
- useTimeEntryForm フック抽出
- DateTimeFields コンポーネント作成

全189テスト合格、Lintエラー0件
